Remove duplicate post modal, misc fixes
