+++ src/app/(community)/layout.tsx
@@ -1,12 +1,15 @@
 "use client"
 
 import Navbar from "@/components/Navbar";
-import CreatePostModal from "@/components/CreatePostModal";
 import React, { useState } from "react";
 import { Toaster } from "@/components/ui/toaster"
 import { AppProgressBar as ProgressBar } from 'next-nprogress-bar';
 import { FOCUS_FONT } from "@/utils/consts";
 import { useUser } from "@/hooks/user";
+import EditPostModal from "@/components/EditPostModal";
+import { Disability } from "@/utils/types/disability";
+import { createPost } from "@/server/db/actions/PostActions";
+import { useToast } from "@/hooks/use-toast";
 
 type CommunityLayoutProps = {
   children: React.ReactNode;
@@ -17,14 +20,52 @@ export default function CommunityLayout({ children }: CommunityLayoutProps) {
   const openCreatePostModal = () => setCreatePostModal(true);
   const closeCreatePostModal = () => setCreatePostModal(false);
   const user = useUser();
+  const { toast } = useToast();
 
   if (!user) return null;
+
+  const notifySuccess = () => {
+    toast({
+      title: "Post successfully added",
+      description: "Your post has been successfully added to the community.",
+    });
+  };
+
+  const notifyFailure = () => {
+    toast({
+      title: "Failed to add post",
+      description: "There was an error adding your post. Please try again.",
+    });
+  };
+
+  async function onPostSubmit(title: string, content: string, tags: Disability[]) {
+    if (!user) return;
+    try {
+      const formattedData = {
+        author: user._id,
+        title,
+        content,
+        tags: tags.map((tag) => tag._id)
+      };
+      await createPost(formattedData);
+      notifySuccess();
+    } catch (error) {
+      notifyFailure();
+      throw error;
+    }
+  }
 
   return (
     <html lang='en'>
       <body className={FOCUS_FONT.className}>
         <Navbar openModal={openCreatePostModal} user={user}/>
-        <CreatePostModal isOpen={isCreatePostModalOpen} openModal={openCreatePostModal} closeModal={closeCreatePostModal} user={user}/>
+        <EditPostModal
+          modalTitle="Create New Post"
+          isOpen={isCreatePostModalOpen}
+          openModal={openCreatePostModal}
+          closeModal={closeCreatePostModal}
+          onSubmit={onPostSubmit}
+        />
         <div className="mx-48 mt-[100px] p-4">
           {children}
           <Toaster />
